Поиск данных/requirements.txt
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.6.1
